# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Faso Atlas API
  version: 1.1.0
  description: |
    Contrat API principal pour les clients web et mobile.
    Base path: /api/v1
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://faso-atlas-api.onrender.com
    description: Production
...
paths:
...
  /api/v1/payments/checkout:
    post:
      tags: [Payments]
      summary: Create a Stripe Checkout session for a reservation
      operationId: createCheckout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    CheckoutRequest:
      type: object
      required: [reservationId]
      properties:
        reservationId:
          type: integer
          format: int64
    CheckoutResponse:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
          format: uri
          description: Stripe Checkout URL to redirect the user to
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true